+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ master (unreleased)
 
 - Added a tox job to update/refresh the swagger-ui related static files (#210 / #213) - including documentation for developers.
 - Remove the field size limit for the model field `formidable.models.Item.label` (#225).
+- handle decimal values in Number fields (#227)
 
 Release 0.10.0 (2017-04-28)
 ===========================